+++ CheckedExceptions/CheckedExceptionsAnalyzer.Delegates.cs
@@ -1,0 +1,69 @@
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+using Microsoft.CodeAnalysis.Diagnostics;
+
+namespace Sundstrom.CheckedExceptions;
+
+partial class CheckedExceptionsAnalyzer
+{
+    /// <summary>
+    /// Resolves the target method symbol from a delegate, lambda, or method group.
+    /// </summary>
+    private static IMethodSymbol GetTargetMethodSymbol(SyntaxNodeAnalysisContext context, InvocationExpressionSyntax invocation)
+    {
+        var expression = invocation.Expression;
+
+        // Get the symbol of the expression being invoked
+        var symbolInfo = context.SemanticModel.GetSymbolInfo(expression);
+        var symbol = symbolInfo.Symbol ?? symbolInfo.CandidateSymbols.FirstOrDefault();
+
+        if (symbol is null)
+            return null;
+
+        if (symbol is ILocalSymbol localSymbol)
+        {
+            // Get the syntax node where the local variable is declared
+            var declaringSyntaxReference = localSymbol.DeclaringSyntaxReferences.FirstOrDefault();
+            if (declaringSyntaxReference is not null)
+            {
+                var syntaxNode = declaringSyntaxReference.GetSyntax();
+
+                if (syntaxNode is VariableDeclaratorSyntax variableDeclarator)
+                {
+                    var initializer = variableDeclarator.Initializer?.Value;
+
+                    if (initializer is not null)
+                    {
+                        // Handle lambdas
+                        if (initializer is AnonymousFunctionExpressionSyntax anonymousFunction)
+                        {
+                            var lambdaSymbol = context.SemanticModel.GetSymbolInfo(anonymousFunction).Symbol as IMethodSymbol;
+                            if (lambdaSymbol is not null)
+                                return lambdaSymbol;
+                        }
+
+                        // Handle method groups
+                        if (initializer is IdentifierNameSyntax || initializer is MemberAccessExpressionSyntax)
+                        {
+                            var methodGroupSymbol = context.SemanticModel.GetSymbolInfo(initializer).Symbol as IMethodSymbol;
+                            if (methodGroupSymbol is not null)
+                                return methodGroupSymbol;
+                        }
+
+                        // Get the method symbol of the initializer (lambda or method group)
+                        var initializerSymbolInfo = context.SemanticModel.GetSymbolInfo(initializer);
+                        var initializerSymbol = initializerSymbolInfo.Symbol ?? initializerSymbolInfo.CandidateSymbols.FirstOrDefault();
+
+                        if (initializerSymbol is IMethodSymbol targetMethodSymbol)
+                        {
+                            return targetMethodSymbol;
+                        }
+                    }
+                }
+            }
+        }
+
+        return null;
+    }
+}
