+++ CheckedExceptions/CheckedExceptionsAnalyzer.XmlDocsHelper.cs
@@ -1,6 +1,4 @@
-using System.Collections.Concurrent;
 using System.Collections.Immutable;
-using System.Xml;
 using System.Xml.Linq;
 
 using Microsoft.CodeAnalysis;
