+++ Test/Test.csproj
@@ -16,19 +16,20 @@
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
   </PropertyGroup> -->
 
-  <!-- <ItemGroup>
+  <ItemGroup>
+    <PackageReference Include="Sundstrom.CheckedExceptions" Version="2.0.2-dev20250811074111">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
     </PackageReference>
-  </ItemGroup> -->
+  </ItemGroup>
 
-  <ItemGroup>
+  <!-- <ItemGroup>
     <ProjectReference Include="..\CheckedExceptions\CheckedExceptions.csproj"
       PrivateAssets="all"
       ReferenceOutputAssembly="true"
       OutputItemType="Analyzer"
       SetTargetFramework="TargetFramework=netstandard2.0" />
-  </ItemGroup>
+  </ItemGroup> -->
 
   <ItemGroup>
     <AdditionalFiles Include="CheckedExceptions.settings.json" />
